+++ doc/changes.rst
@@ -5,7 +5,9 @@ surveysim change log
 0.9.2 (unreleased)
 ------------------
 
-* No changes yet.
+* Replay historical Mayall daily weather.
+* Implement partial-night dome closure.
+* Requires desimodel >= 0.9.8 and desisurvey >= 0.10.4.
 
 0.9.1 (2018-06-27)
 ------------------